Refactor training code and remove unused / 2-class code, files and directories
