--- deepminer/.gitignore
@@ -1,1 +1,0 @@
-output
