--- deepminer/pretrained
@@ -1,1 +1,0 @@
-../training/pretrained/
