+++ training/analyze_full_images.py
@@ -8,95 +8,20 @@ import hashlib
 import numpy as np
 import torch
 import torch.utils.data
-import torch.backends.cudnn as cudnn
 import torch.nn as nn
-import torchvision.transforms as transforms
 from munch import Munch
 from torch.autograd import Variable
 from tqdm import tqdm as tqdm
-from PIL import Image, ImageOps
 
-import models.resnet
+from dataset import DDSM
+from models.resnet_3class import get_resnet152_3class_model
 from db.database import DB
 
 
-class DDSM(torch.utils.data.Dataset):
-    def __init__(self, root, image_list_path, target_size, transform):
-        self.root = root
-        name2class = {
-            'normal': 0,
-            'benign': 1,
-            'cancer': 2,
-        }
-        with open(image_list_path, 'r') as f:
-            self.images = [(line.strip(), name2class[line.strip()[:6]]) for line in f.readlines()]
-        self.image_names = [filename for filename, ground_truth in self.images]
-        self.target_size = target_size
-        self.transform = transform
-
-        weight = np.unique([label for _, label in self.images], return_counts=True)[1]
-        self.weight = 1 / (weight / np.amin(weight))
-
-        print(np.unique([ground_truth for filename, ground_truth in self.images], return_counts=True))
-
-    def __len__(self):
-        return len(self.images)
-
-    def __getitem__(self, idx):
-        image_name, ground_truth = self.images[idx]
-        image = Image.open(os.path.join(self.root, image_name))
-        min_dim = min(image.size)
-        ratio = self.target_size / min_dim
-        new_size = (int(ratio * image.size[0]), int(ratio * image.size[1]))
-        image = image.resize(new_size, resample=Image.BILINEAR)  # image shape is now (~1500, 896)
-        delta_w = self.target_size - new_size[0]
-        delta_h = self.target_size - new_size[1]
-        padding = (delta_w // 2, delta_h // 2, delta_w - (delta_w // 2), delta_h - (delta_h // 2))
-        image = ImageOps.expand(image, padding)
-        image = np.asarray(image)
-        image = np.broadcast_to(np.expand_dims(image, 2), image.shape + (3,))  # image shape is now (~1500, 896, 3)
-        image = self.transform(image)  # image shape is now (3, ~1500, 896) and a it is a tensor
-        return image, ground_truth
-
-
-def get_dataset():
-    normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
-    target_size = 512
-    raw_image_dir = '../data/ddsm_raw'
-    image_list_val = '../data/ddsm_raw_image_lists/val.txt'
-
-    val_dataset = DDSM(raw_image_dir, image_list_val, target_size, transforms.Compose([
-        transforms.ToTensor(),
-        normalize,
-    ]))
-
-    return val_dataset
-
-
 def prepare_model(cfg):
-    print("=> creating model '{}'".format(cfg.arch.model))
-    if not cfg.arch.model == 'resnet152':
-        raise KeyError("Only ResNet152 is supported, not %s" % cfg.arch.model)
-
-    model = models.resnet.resnet152(pretrained=cfg.arch.pretrained)
-    model.fc = nn.Linear(2048, cfg.arch.num_classes)
-    features_layer = model.layer4
-
-    model = torch.nn.DataParallel(model).cuda()
-    cudnn.benchmark = True
-
     resume_path = cfg.training.resume.replace(cfg.training.resume[-16:-8], '{:08}'.format(args.epoch))
     resume_path = os.path.join('../training', resume_path)
-    if os.path.isfile(resume_path):
-        print("=> loading checkpoint '{}'".format(resume_path))
-        checkpoint = torch.load(resume_path)
-        model.load_state_dict(checkpoint['state_dict'])
-        model.eval()
-        print("=> loaded checkpoint '{}' (epoch {})".format(resume_path, checkpoint['epoch']))
-    else:
-        print("=> no checkpoint found at '{}'".format(resume_path))
-
-    # convert_fc_to_conv(model)
+    model, epoch, optimizer_state, features_layer = get_resnet152_3class_model(resume_path)
     return model, features_layer, resume_path
 
 
@@ -231,7 +156,7 @@ def print_statistics(ranked_units, max_activation_per_unit_per_input):
 
 def analyze_full_images(args, cfg, db_path):
     model, features_layer, checkpoint_path = prepare_model(cfg)
-    val_dataset = get_dataset()
+    val_dataset = DDSM.create_full_image_dataset('val')
 
     max_activation_per_unit_per_input, classifications = run_model_on_all_images(model, features_layer, val_dataset)
     unit_id_and_count_per_class, ranked_units, weighted_max_activations = create_unit_ranking(model, max_activation_per_unit_per_input)
