--- training/config/2class/inception_v3.yml
@@ -1,23 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: inception_v3
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: []
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    logs_dir: logs
-    checkpoints_dir: checkpoints
-    experiment_name: inception_v3
-    print_freq: 10
-    checkpoint_epochs: 1
-    log_dir:
-    epochs: 100
-    resume:
