--- training/config/2class/resnet152.yml
@@ -1,23 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: resnet152
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: [4]
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    logs_dir: logs
-    checkpoints_dir: checkpoints
-    experiment_name: resnet152
-    print_freq: 10
-    checkpoint_epochs: 1
-    log_dir:
-    epochs: 100
-    resume:
