+++ training/config/resnet152_full_images.yml
