--- training/download_pretrained.sh
@@ -1,6 +1,0 @@
-cd pretrained
-wget http://data.csail.mit.edu/places/medical/pretrained/2class/alexnet/checkpoint_00000045.pth.tar -P alexnet
-wget http://data.csail.mit.edu/places/medical/pretrained/2class/vgg16/checkpoint_00000012.pth.tar -P vgg16
-wget http://data.csail.mit.edu/places/medical/pretrained/2class/inception_v3/checkpoint_00000007.pth.tar -P inception_v3
-wget http://data.csail.mit.edu/places/medical/pretrained/2class/resnet152/checkpoint_00000005.pth.tar -P resnet152
-wget http://data.csail.mit.edu/places/medical/pretrained/3class/resnet152/checkpoint_00000005.pth.tar -P resnet152_3class
