--- training/pretrained/alexnet/config.yml
@@ -1,17 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: alexnet
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: []
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    experiment_name: alexnet
-    resume: pretrained/alexnet/checkpoint_00000045.pth.tar
