--- training/pretrained/inception_v3/config.yml
@@ -1,17 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: inception_v3
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: []
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    experiment_name: inception_v3
-    resume: pretrained/inception_v3/checkpoint_00000007.pth.tar
