--- training/pretrained/resnet152/config.yml
@@ -1,17 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: resnet152
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: [4]
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    experiment_name: resnet152
-    resume: pretrained/resnet152/checkpoint_00000005.pth.tar
