--- training/pretrained/vgg16/config.yml
@@ -1,17 +1,0 @@
-data:
-    root: ../data/ddsm_2class
-    batch_size: 32
-    workers: 4
-arch:
-    num_classes: 2
-    model: vgg16
-    pretrained: true
-optimizer:
-    type: sgd
-    lr: 0.0001
-    lr_decay_epochs: []
-    momentum: 0.9
-    weight_decay: 0.0001
-training:
-    experiment_name: vgg16
-    resume: pretrained/vgg16/checkpoint_00000012.pth.tar
